requirements.txt
pandas==2.1.4
scikit_learn==1.3.2
tqdm==4.65.0
adapter-transformers
tabulate==0.9.0
datasets